[ICU-900] Fixed some compiler warnings. X-Commit-URL: https://ssl.icu-project.org/trac/changeset/8400 Former-commit-id: e3a7e70d8903f065760d9d981628f8a6d93608f7

+++ icu4c/source/i18n/usearch.cpp
@@ -431,7 +431,7 @@ inline UBool isBreakUnit(const UStringSearch *strsrch, int32_t start,
                     return FALSE;
                 }
             }
-            if (ucol_next(coleiter, &status) != UCOL_NULLORDER) {
+            if (ucol_next(coleiter, &status) != (int32_t)UCOL_NULLORDER) {
                 // extra collation elements at the end of the match
                 return FALSE;
             }
